+++ src/interceptors/model-validation/BaseRefExistsInterceptor.ts
