+++ src/interceptors/model-validation/ExistenceInterceptorOpts.ts
